requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic==1.10.11
python-multipart
passlib[bcrypt]
python-jose